xbl_log: Add driver to read secondary bootloader (XBL) log
Secondary bootloader (XBL) boot log holds information to identify
various firmware configuration currently set on the SoC. The XBL
log is stored in a predefined reserved memory region.

This drivers provides a way to print XBL logs on the console. To
do so, it provides a debugfs entry which captures the logs stored
in this reserved memory region. This entry can now be used to read
and print the XBL logs to console.

User can use the below command to print XBL log to console:
        cat /sys/kernel/debug/xbl_log

Change-Id: Ie95c9e9d9caa7895be83c0f10aa8b06200f1c1f5
Signed-off-by: Ninad Naik <quic_ninanaik@quicinc.com>